improved list function properties

+++ plotmathfunctions/2d/2d-library.js
@@ -1,4 +1,8 @@
-function listFunctions() {
+function listFunctions(showDeleteButton,showAddNewButton) {
+
+  showDeleteButton=(showDeleteButton==null?true:showDeleteButton);
+  showAddNewButton=(showAddNewButton==null?true:showAddNewButton);
+
   var strTable = "<table>";
   var i=0;
 
@@ -11,7 +15,9 @@ function listFunctions() {
           strTable += "<td>" + key + "</td>";
           strTable += `<td><input type='text' id="${key}" value="${params.functions[key]}"></td>`; //<td> " + params.functions[key] + "</td>";
           strTable += "<td>";
+          if (showDeleteButton){
           strTable += `<a href="#" onclick="return deleteFunction('${key}')"> delete </a> `;
+          }
           strTable += `<a href="#" onclick="return updateFunction('${key}')"> update </a> `;
           strTable += "</td>";
           strTable += "</tr>";
@@ -19,14 +25,16 @@ function listFunctions() {
       }
     }
   }
-  //add new function 
-  i++;
-  strTable += "<tr>";
-          strTable += "<td> f" + i + "</td>";
-          strTable += "<td><input type='text' id='newF'></td>";
-          strTable += "<td>" + `<a href="#" onclick="addFunction('f${i}')"> add </a>` + "</td>";
-          strTable += "</tr>";
+  if (showAddNewButton){
 
+    //add new function 
+    i++;
+    strTable += "<tr>";
+            strTable += "<td> f" + i + "</td>";
+            strTable += "<td><input type='text' id='newF'></td>";
+            strTable += "<td>" + `<a href="#" onclick="addFunction('f${i}')"> add </a>` + "</td>";
+            strTable += "</tr>";
+  } //end if
   strTable += "</table>";
 
   return strTable;
